+++ libs/infinity_emb/infinity_emb/__init__.py
@@ -1,7 +1,8 @@
-__all__ = ["logger", "create_server", "inference", "fastapi_schemas", "__version__"]
+__all__ = ["transformer", "inference", "fastapi_schemas","logger", "create_server", "__version__"]
 import importlib.metadata
 
-from infinity_emb import fastapi_schemas, inference
+from infinity_emb import fastapi_schemas, inference, transformer
+# reexports
 from infinity_emb.infinity_server import create_server
 from infinity_emb.log_handler import logger
