+++ libs/infinity_emb/pyproject.toml
@@ -8,7 +8,7 @@ readme = "README.md"
 packages = [{include = "infinity_emb"}]
 
 [tool.poetry.dependencies]
-python = ">=3.10,<4.0"
+python = ">=3.10,<4"
 fastapi = "^0.103.2"
 pydantic = ">=2.4.2,<3"
 torch = ">=2.0.0, !=2.0.1, !=2.1.0"
