+++ libs/infinity_emb/tests/end_to_end/test_sentence_transformers.py
@@ -10,7 +10,7 @@ from httpx import AsyncClient
 from sentence_transformers import SentenceTransformer  # type: ignore
 
 from infinity_emb import create_server
-from infinity_emb.inference.models import InferenceEngine
+from infinity_emb.transformer.utils import InferenceEngine
 
 PREFIX = "/v1_ct2"
 model = pytest.DEFAULT_BERT_MODEL
